+++ packages/expo-brownfield/gradle-plugins/publish/src/main/kotlin/expo/modules/plugin/utils.kt
@@ -183,7 +183,7 @@ internal fun getHermesVersion(project: Project): String {
  */
 internal fun getHermesVersionInternal(project: Project): String {
   val process =
-    ProcessBuilder("node", "--print", "require('hermes-compiler/package.json').version")
+    ProcessBuilder("node", "--print", "require(require.resolve('hermes-compiler/package.json', { paths: [require.resolve('react-native/package.json')] })).version")
       .directory(project.rootProject.projectDir)
       .redirectErrorStream(true)
       .start()